requirements.txt
amqp==2.5.1
asgi-redis==1.4.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
backports.csv==1.0.7
billiard==3.6.1.0
boto3==1.9.153
botocore==1.12.153
celery==4.3.0
certifi==2019.3.9
cffi==1.13.1
channels==2.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-allauth==0.39.1
django-appconf==1.0.3
django-channels==0.7.0
django-ckeditor==5.7.1
django-colorfield==0.1.15
django-colorful==1.3
django-cors-headers==3.2.1
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-js-asset==1.2.2
django-select2==7.1.0
django-suit==0.2.26
djangorestframework==3.10.2
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
docutils==0.14
et-xmlfile==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
jdcal==1.4.1
jmespath==0.9.4
kombu==4.6.5
Markdown==3.1.1
more-itertools==7.2.0
msgpack-python==0.5.6
ndg-httpsclient==0.5.1
nltk==3.4.5
numpy==1.17.1
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pipdeptree==0.13.2
pipreqs==0.4.9
protobuf==3.9.1
psycopg2-binary==2.8.4
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
stanfordnlp==0.2.0
tablib==0.13.0
torch==1.2.0
tqdm==4.35.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.24.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
zipp==0.6.0
zope.interface==4.6.0
